# public/api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "POS API",
        "description": "REST API untuk Sistem Point of Sale (POS)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8303/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/pelanggans": {
            "get": {
                "tags": [
                    "Pelanggans"
                ],
                "summary": "Get all pelanggan with pagination.",
                "operationId": "99125c2c4f14377d85456baa7be87c80",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pagination"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pelanggans"
                ],
                "summary": "Create new pelanggan.",
                "operationId": "9bbaf5f4a0dc1e778507adae6910185c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PelangganCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pelanggans/{pelanggan}": {
            "get": {
                "tags": [
                    "Pelanggans"
                ],
                "summary": "Get single pelanggan detail.",
                "operationId": "a34228f32979be24cde08c7d5cea1b0e",
                "parameters": [
                    {
                        "name": "pelanggan",
                        "in": "path",
                        "description": "ID pelanggan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pelanggans"
                ],
                "summary": "Update pelanggan.",
                "operationId": "57f6ac416a90de5133d7b36c34c7d1dd",
                "parameters": [
                    {
                        "name": "pelanggan",
                        "in": "path",
                        "description": "ID pelanggan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "domisili": {
                                        "type": "string"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "enum": [
                                            "PRIA",
                                            "WANITA"
                                        ]
                                    },
                                    "poin": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pelanggans"
                ],
                "summary": "Delete pelanggan.",
                "operationId": "53cadc43da8bc8d690b0a54fd05acf08",
                "parameters": [
                    {
                        "name": "pelanggan",
                        "in": "path",
                        "description": "ID pelanggan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PelangganCreate": {
                "required": [
                    "nama",
                    "jenis_kelamin"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Pagination"
                    },
                    {
                        "properties": {
                            "id_pelanggan": {
                                "type": "string",
                                "example": "PGN123"
                            },
                            "nama": {
                                "type": "string",
                                "example": "John Doe"
                            },
                            "domisili": {
                                "type": "string",
                                "example": "Jakarta"
                            },
                            "jenis_kelamin": {
                                "type": "string",
                                "enum": [
                                    "PRIA",
                                    "WANITA"
                                ],
                                "example": "PRIA"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "ApiResponse": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "first_page_url": {
                        "type": "string"
                    },
                    "from": {
                        "type": "integer",
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "last_page_url": {
                        "type": "string"
                    },
                    "next_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "prev_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "to": {
                        "type": "integer",
                        "nullable": true
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}